requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
markdown==3.4.4
python-dotenv==1.0.1